+++ src/command.rs
@@ -150,15 +150,15 @@ fn handle_from_toml_error(error: &FromTomlError) -> ErrorHandling {
             gitz_version, ..
         } => {
             error!("{error}.");
-            hint!(
-                "\n{}\n{}\n{}\n{}\n{}\n{}",
-                format_args!("Your {CONFIG_FILE_NAME} has been created by a development version of git-z."),
-                "However, configurations produced by a development version are only",
-                "supported by the immediately following release.\n",
-                format_args!("To update from this version, you can install git-z {gitz_version},"),
-                "run `git z update`, then update to the latest version and run",
-                "`git z update` again."
-            );
+            hint! {"
+                Your {CONFIG_FILE_NAME} has been created by a development version of git-z.
+                However, configurations produced by a development version are only
+                supported by the immediately following release.
+
+                To update from this version, you can install git-z {gitz_version}
+                run `git z update`, then update to the latest version and run
+                `git z update` again.\
+            "};
         }
         FromTomlError::ParseError(parse_error) => {
             error!("Invalid configuration in {CONFIG_FILE_NAME}.");
@@ -208,15 +208,15 @@ fn handle_update_error(error: &UpdateError) -> ErrorHandling {
         }
         UpdateError::UnsupportedDevelopmentVersion { gitz_version, .. } => {
             error!("{error}.");
-            hint!(
-                "\n{}\n{}\n{}\n{}\n{}\n{}",
-                "`git z update` can update a configuration from any previous release.",
-                "However, configurations produced by a development version can only be",
-                "updated by the immediately following release.\n",
-                format_args!("To update from this version, you can install git-z {gitz_version},"),
-                "run `git z update`, then update to the latest version and run",
-                "`git z update` again."
-            );
+            hint! {"
+                `git z update` can update a configuration from any previous release.
+                However, configurations produced by a development version can only be
+                updated by the immediately following release.
+
+                To update from this version, you can install git-z {gitz_version},
+                run `git z update`, then update to the latest version and run
+                `git z update` again.\
+            "};
         }
     }